requirements.txt

pbr
python-etcd
